Improvements to the design and global connection function.
The app will open maximized by default.

When commands are running, you won't be able to interact with the rest of the app.

Instead of a "Please wait" text, the app will show an "infinite progress bar."

There's now an option to connect or disconnect a plug in all apps that have the selected plug. This section doesn't include all existing plugs.
When I tested this feature by selecting multiple plugs, the app would freeze, so this section is limited to one plug at a time.

When the QT Oxygen style is available, it will be used by default, but you can simply change the name "Oxygen" to whatever you like.