Version 0.3
Fixed an app icon that wasn't being used.

Added the Spanish language.

The area where the progress bar was located has been removed; the notifications that appeared there will now appear in the title bar.

The resizing status of the apps and plugins area will be remembered.

The radio button for the global connection and global disconnection options will no longer be deselected after the commands are completed.

Removed unnecessary imports.